+++ io-core/src/jvmTest/kotlin/io/github/fukusaka/keel/io/PoolAllocatorTest.jvm.kt
@@ -1,0 +1,6 @@
+package io.github.fukusaka.keel.io
+
+actual fun createPoolAllocator(bufferSize: Int, maxPoolSize: Int): BufferAllocator =
+    PooledDirectAllocator(bufferSize, maxPoolSize)
+
+actual fun isPoolAllocator(): Boolean = true